Build 404 page for static sites
Both Netlify and Vercel default to reading a static site's 404 page
from a file called `404.html` in the root of the build folder.

I'm therefore considering this as a convention, and have hard coded the
path.

Note that the log message that announces that a page has been rendered
is now guarded by a check to see whether the file contents have changed.

We don't know which template (or app code) the 404.html page was created
from, so we're unable to check whether or not it's out of date. So we
re-generate 404.html on every run. Printing a message to say that we've
done so even if the source files haven't changed could be confusing.
Hence the check to see if the contents have changed.

+++ lib/nesta/static/site_content.rb
@@ -9,6 +9,7 @@ module Nesta
         @build_dir = build_dir
         @logger = logger
         @app = Nesta::App.new
+        set_app_root
       end
 
       def log(message)
@@ -27,23 +28,29 @@ module Nesta
       end
 
       def render_pages
-        set_app_root
         Nesta::Page.find_all.each do |page|
           target = HtmlFile.new(@build_dir, page).filename
           source = page.filename
           task = Rake::FileTask.define_task(target => source) do
-            http_code, markup = render_page(page, target)
-            save_markup(target, markup)
+            save_markup(target, render_page(page.abspath, source, target))
           end
           task.invoke
         end
       end
 
-      def rack_environment(page)
+      def render_not_found
+        path_info = '/404'
+        source = 'no-such-file.md'
+        target = File.join(@build_dir, '404.html')
+        markup = render_page(path_info, source, target, expected_code: 404)
+        save_markup(target, markup)
+      end
+
+      def rack_environment(abspath)
         {
           'REQUEST_METHOD' => 'GET',
           'SCRIPT_NAME' => '',
-          'PATH_INFO' => page.abspath,
+          'PATH_INFO' => abspath,
           'QUERY_STRING' => '',
           'SERVER_NAME' => 'localhost',
           'SERVER_PROTOCOL' => 'https',
@@ -53,18 +60,20 @@ module Nesta
         }
       end
 
-      def render_page(page, html_path)
-        http_code, headers, body = @app.call(rack_environment(page))
-        if http_code != 200
-          raise RuntimeError, "Can't render #{html_path} from #{page.filename}"
+      def render_page(abspath, content_path, html_path, expected_code: 200)
+        http_code, headers, body = @app.call(rack_environment(abspath))
+        if http_code != expected_code
+          raise RuntimeError, "Can't render #{html_path} from #{content_path}"
         end
-        log("Rendered #{html_path}: #{http_code}")
-        [http_code, body.join]
+        body.join
       end
 
       def save_markup(filename, content)
         FileUtils.mkdir_p(File.dirname(filename))
-        open(filename, 'w') { |output| output.puts(content) }
+        if (! File.exist?(filename)) || (open(filename, 'r').read != content)
+          open(filename, 'w') { |output| output.puts(content) }
+          log("Rendered #{filename}")
+        end
       end
     end
   end
